Move JS controllers into a separate dir
Move controllers into a separate dir to make it easier to see which
modules are the entry points for site bundles.
