+++ h/static/scripts/admin-site.js
@@ -9,8 +9,8 @@ if (settings.raven) {
 window.$ = window.jQuery = require('jquery');
 require('bootstrap');
 
-var AdminUsersController = require('./admin-users');
-var upgradeElements = require('./upgrade-elements');
+var AdminUsersController = require('./controllers/admin-users-controller');
+var upgradeElements = require('./controllers/upgrade-elements');
 
 var controllers = {
   '.js-users-delete-form': AdminUsersController,
