+++ h/static/scripts/controllers/upgrade-elements.js
